feat(api-headless-cms): create model with fields (#2443)
