+++ packages/api-headless-cms/src/content/plugins/crud/contentModel/createFieldModels.ts
@@ -1,12 +1,12 @@
-import { CmsModelField, CmsModel, CmsModelUpdateInput } from "~/types";
+import { CmsModelField, CmsModelFieldInput } from "~/types";
 import { ContentModelFieldModel } from "./models";
 
-export const createFieldModels = async (
-    _: CmsModel,
-    data: CmsModelUpdateInput
-): Promise<CmsModelField[]> => {
+export const createFieldModels = async (input?: CmsModelFieldInput[]): Promise<CmsModelField[]> => {
+    if (!input || input.length === 0) {
+        return [];
+    }
     const fields: CmsModelField[] = [];
-    for (const field of data.fields) {
+    for (const field of input) {
         const fieldData = new ContentModelFieldModel().populate(field);
         await fieldData.validate();
         fields.push(await fieldData.toJSON());
